过大页面基本信息插入数据库；入库插入方法封装

+++ src/main/java/com/trs/gov/kpi/entity/responsedata/PageSpaceResponse.java
@@ -16,9 +16,9 @@ public class PageSpaceResponse {
 
     private String pageLink;
 
-    private Integer replySpeed;
+    private Long replySpeed;
 
-    private Integer pageSpace;
+    private Long pageSpace;
 
     private Date checkTime;
 }
